+++ greynoise-enrichment/source/greynoise_anomali_enrichment.py
@@ -1,22 +1,40 @@
+import copy
+import datetime
+import json
 import os
 import sys
 
 import requests
-
 from AnomaliEnrichment import (
     AnomaliEnrichment,
-    TextWidget,
-    TableWidget,
+    ChartWidget,
+    CompositeItem,
+    HorizontalRuleWidget,
     ItemInWidget,
     ItemTypes,
-    CompositeItem,
+    TableWidget,
+    TextWidget,
 )
 
 api_base = "https://api.greynoise.io/"
 api_key = None
 api_type = "enterprise"
 
-VERSION = "2.3.0"
+VERSION = "2.4.0"
+
+xrange_template_dict = {
+    "chart": {"type": "xrange", "height": 350},
+    "title": {"text": "IP Timeline"},
+    "xAxis": {"type": "datetime"},
+    "yAxis": {
+        "title": {"text": ""},
+        "categories": ["Classification", "rDNS", "ASN"],
+        "gridLineWidth": 0,
+        "reversed": True,
+    },
+    "tooltip": {"pointFormat": "{point.value}"},
+    "series": [{"name": "Timeline", "borderColor": "white", "pointWidth": 25, "data": []}],
+}
 
 
 def enrichIP(anomali_enrichment, search_string):  # noqa: C901
@@ -77,7 +95,7 @@ def enrichIP(anomali_enrichment, search_string):  # noqa: C901
                     TextWidget(
                         ItemInWidget(
                             ItemTypes.Link,
-                            "https://www.greynoise.io/viz/ip/%s" % search_string,
+                            "https://viz.greynoise.io/ip/%s" % search_string,
                             "View on GreyNoise Visualizer",
                         ),
                         True,
@@ -152,10 +170,14 @@ def enrichIP(anomali_enrichment, search_string):  # noqa: C901
                         ItemInWidget(itemValue=response_json.get("name")),
                     ]
                 )
+                date_object = datetime.datetime.strptime(
+                    response_json.get("last_seen", "0000-00-00"), "%Y-%m-%d"
+                ).date()
+                last_seen_formatted = date_object.strftime("%-d %b %Y")
                 table_widget.addRowOfItems(
                     [
                         ItemInWidget(itemValue="Last Seen"),
-                        ItemInWidget(itemValue=response_json.get("last_seen")),
+                        ItemInWidget(itemValue=last_seen_formatted),
                     ]
                 )
 
@@ -196,8 +218,8 @@ def enrichIP(anomali_enrichment, search_string):  # noqa: C901
                 TextWidget(
                     ItemInWidget(
                         ItemTypes.String,
-                        "GreyNoise Info for %s" % search_string,
-                        "GreyNoise Info for %s" % search_string,
+                        "GreyNoise Intel for %s" % search_string,
+                        "GreyNoise Intel for %s" % search_string,
                         "#A9A9A9",
                         "#FFFFFF",
                         "30px",
@@ -210,7 +232,7 @@ def enrichIP(anomali_enrichment, search_string):  # noqa: C901
                 TextWidget(
                     ItemInWidget(
                         ItemTypes.Link,
-                        "https://www.greynoise.io/viz/ip/%s" % search_string,
+                        "https://viz.greynoise.io/ip/%s" % search_string,
                         "View on GreyNoise Visualizer",
                     ),
                     True,
@@ -234,16 +256,20 @@ def enrichIP(anomali_enrichment, search_string):  # noqa: C901
 
             # Table Widget #1 Start
             table_widget = TableWidget("Details", ["Key", "Value"], columnWidths=["20%", "80%"])
+            date_object = datetime.datetime.strptime(response_json.get("last_seen", "0000-00-00"), "%Y-%m-%d").date()
+            last_seen_formatted = date_object.strftime("%-d %b %Y")
             table_widget.addRowOfItems(
                 [
                     ItemInWidget(itemValue="Last Seen"),
-                    ItemInWidget(itemValue=response_json["last_seen"]),
+                    ItemInWidget(itemValue=last_seen_formatted),
                 ]
             )
+            date_object = datetime.datetime.strptime(response_json.get("first_seen", "0000-00-00"), "%Y-%m-%d").date()
+            first_seen_formatted = date_object.strftime("%-d %b %Y")
             table_widget.addRowOfItems(
                 [
                     ItemInWidget(itemValue="First Seen"),
-                    ItemInWidget(itemValue=response_json["first_seen"]),
+                    ItemInWidget(itemValue=first_seen_formatted),
                 ]
             )
             if response_json["classification"] == "malicious":
@@ -309,7 +335,7 @@ def enrichIP(anomali_enrichment, search_string):  # noqa: C901
             if response_json["metadata"].get("country") and response_json["metadata"].get("country") != "unknown":
                 table_widget_metadata.addRowOfItems(
                     [
-                        ItemInWidget(itemValue="Country"),
+                        ItemInWidget(itemValue="Source Country"),
                         ItemInWidget(itemValue=response_json["metadata"].get("country", "Unknown") or "Unknown"),
                     ]
                 )
@@ -319,8 +345,24 @@ def enrichIP(anomali_enrichment, search_string):  # noqa: C901
             ):
                 table_widget_metadata.addRowOfItems(
                     [
-                        ItemInWidget(itemValue="Country Code"),
+                        ItemInWidget(itemValue="Source Country Code"),
                         ItemInWidget(itemValue=response_json["metadata"].get("country_code", "Unknown") or "Unknown"),
+                    ]
+                )
+            if response_json["metadata"].get("destination_countries"):
+                dest_countries = ", ".join(response_json["metadata"]["destination_countries"])
+                table_widget_metadata.addRowOfItems(
+                    [
+                        ItemInWidget(itemValue="Destination Countries"),
+                        ItemInWidget(itemValue=dest_countries or "Unknown"),
+                    ]
+                )
+            if response_json["metadata"].get("destination_country_codes"):
+                dest_country_codes = ", ".join(response_json["metadata"]["destination_country_codes"])
+                table_widget_metadata.addRowOfItems(
+                    [
+                        ItemInWidget(itemValue="Destination Country Codes"),
+                        ItemInWidget(itemValue=dest_country_codes or "Unknown"),
                     ]
                 )
             if response_json["metadata"].get("region") and response_json["metadata"].get("region") != "unknown":
@@ -631,7 +673,7 @@ def enrichIP(anomali_enrichment, search_string):  # noqa: C901
                 TextWidget(
                     ItemInWidget(
                         ItemTypes.Link,
-                        "https://www.greynoise.io/viz/ip/%s" % search_string,
+                        "https://viz.greynoise.io/ip/%s" % search_string,
                         "View on GreyNoise Visualizer",
                     ),
                     True,
@@ -640,10 +682,14 @@ def enrichIP(anomali_enrichment, search_string):  # noqa: C901
 
             # Table Widget #1 Start
             table_widget = TableWidget("Details", ["Key", "Value"], columnWidths=["20%", "80%"])
+            date_object = datetime.datetime.strptime(
+                riot_response_json.get("last_updated", "0000-00-00T00:00:00Z").split("T")[0], "%Y-%m-%d"
+            ).date()
+            last_updated_formatted = date_object.strftime("%-d %b %Y")
             table_widget.addRowOfItems(
                 [
                     ItemInWidget(itemValue="Last Updated"),
-                    ItemInWidget(itemValue=riot_response_json["last_updated"]),
+                    ItemInWidget(itemValue=last_updated_formatted),
                 ]
             )
             table_widget.addRowOfItems(
@@ -737,21 +783,366 @@ def enrichIP(anomali_enrichment, search_string):  # noqa: C901
             )
         elif response.status_code == 401:
             anomali_enrichment.addException("API Key is Missing, Expired or Incorrect, please verify")
+            anomali_enrichment.addMessage("ERROR", "API Key is Missing, Expired or Incorrect, please verify")
         elif response.status_code == 429:
             anomali_enrichment.addException("API Rate-Limit Reached, Please try again tomorrow.")
+            anomali_enrichment.addMessage("ERROR", "API Rate-Limit Reached, Please try again tomorrow.")
         elif response.status_code == 500:
             anomali_enrichment.addException(
-                "An error occurred with the GreyNoise API, please contact GreyNoise for assistance."
+                "An error occurred with the GreyNoise API, please contact GreyNoise for " "assistance."
+            )
+            anomali_enrichment.addMessage(
+                "ERROR", "An error occurred with the GreyNoise API, please contact " "GreyNoise for assistance. "
             )
     except:  # noqa E722
         anomali_enrichment.addException(
             "enrichIP Unknown Error:%sType: %s%sValue:%s"
             % (os.linesep, sys.exc_info()[0], os.linesep, sys.exc_info()[1])
         )
+        anomali_enrichment.addMessage(
+            "ERROR", "An Unexpected error occurred, please try again. If the error persists please contact support."
+        )
     return anomali_enrichment
 
 
-functions = {"enrichIP": enrichIP}
+def enrichIPSim(anomali_enrichment, search_string):  # noqa: C901
+    try:
+        if api_type.lower() == "community":
+            anomali_enrichment.addMessage("ERROR", "IP Similarity Not Supported with Community API Key")
+        else:
+            # builds response if paid api is being used
+            similarity_response = requests.get(
+                api_base + "v3/similarity/ips/" + search_string + "?limit=50&minimum_score=0.9",
+                headers={
+                    "Accept": "application/json",
+                    "key": api_key,
+                    "User-Agent": "greynoise-anomali-enrichment-" + VERSION,
+                },
+            )
+            similarity_response_json = similarity_response.json()
+
+            if (
+                similarity_response.status_code == 200
+                and similarity_response_json
+                and int(similarity_response_json["total"]) > 1
+            ):
+                anomali_enrichment.addWidget(
+                    TextWidget(
+                        ItemInWidget(
+                            ItemTypes.String,
+                            "GreyNoise Similarity Intel for %s" % search_string,
+                            "GreyNoise Similarity Intel for %s" % search_string,
+                            "#A9A9A9",
+                            "#FFFFFF",
+                            "30px",
+                            "bold",
+                        ),
+                        True,
+                    )
+                )
+                if int(similarity_response_json["total"]) > 50:
+                    anomali_enrichment.addWidget(
+                        TextWidget(
+                            ItemInWidget(
+                                ItemTypes.String,
+                                "Showing first 50 IPs of {} IPs that have a similarity score "
+                                "of 90% or above to {}".format(similarity_response_json["total"], search_string),
+                            )
+                        )
+                    )
+                else:
+                    anomali_enrichment.addWidget(
+                        TextWidget(
+                            ItemInWidget(
+                                ItemTypes.String,
+                                "Showing {} IPs that have a similarity score "
+                                "of 90% or above to {}".format(similarity_response_json["total"], search_string),
+                            )
+                        )
+                    )
+                anomali_enrichment.addWidget(
+                    TextWidget(
+                        ItemInWidget(
+                            ItemTypes.Link,
+                            "https://viz.greynoise.io/ip-similarity/%s" % search_string,
+                            "View IP Similarity UI on GreyNoise Visualizer",
+                        ),
+                        True,
+                    )
+                )
+                # Table Widget #4 Start
+                table_widget_similarity = TableWidget(
+                    "Similar IPs",
+                    [
+                        "IP",
+                        "Score",
+                        "Classification",
+                        "Actor",
+                        "Last Seen",
+                        "Organization",
+                        "Features Matched",
+                    ],
+                    columnWidths=["10%", "5%", "10%", "15%", "10%", "15%", "35%"],
+                )
+                for similar_ip in similarity_response_json["similar_ips"]:
+                    features = ", ".join(similar_ip.get("features", None))
+                    similar_ip["score"] = str(int(similar_ip.get("score", "0") * 100)) + "%"
+                    date_object = datetime.datetime.strptime(
+                        similar_ip.get("last_seen", "0000-00-00"), "%Y-%m-%d"
+                    ).date()
+                    last_seen_formatted = date_object.strftime("%-d %b %Y")
+                    table_widget_similarity.addRowOfItems(
+                        [
+                            ItemInWidget(
+                                ItemTypes.Link,
+                                "/detail/v2/ip?value={}".format(similar_ip.get("ip")),
+                                similar_ip.get("ip"),
+                            ),
+                            ItemInWidget(itemValue=similar_ip.get("score", None)),
+                            ItemInWidget(itemValue=similar_ip.get("classification", None)),
+                            ItemInWidget(itemValue=similar_ip.get("actor", None)),
+                            ItemInWidget(itemValue=last_seen_formatted),
+                            ItemInWidget(itemValue=similar_ip.get("organization", None)),
+                            ItemInWidget(itemValue=features),
+                        ]
+                    )
+                anomali_enrichment.addWidget(table_widget_similarity)
+                # Table Widget #4 End
+            elif similarity_response.status_code == 403:
+                horizontal_rule_widget = HorizontalRuleWidget()
+                anomali_enrichment.addWidget(horizontal_rule_widget)
+                anomali_enrichment.addWidget(
+                    TextWidget(
+                        ItemInWidget(
+                            ItemTypes.String,
+                            "GreyNoise Similarity Intel for %s is not available with your Subscription" % search_string,
+                            "GreyNoise Similarity Intel for %s is not available with your Subscription" % search_string,
+                            "#A9A9A9",
+                            "#FFFFFF",
+                            "30px",
+                            "bold",
+                        ),
+                        True,
+                    )
+                )
+            elif similarity_response.status_code == 401:
+                anomali_enrichment.addMessage("ERROR", "API Key is Missing, Expired or Incorrect, please verify")
+            elif similarity_response.status_code == 429:
+                anomali_enrichment.addMessage("ERROR", "API Rate-Limit Reached, Please try again tomorrow.")
+            elif similarity_response.status_code == 500:
+                anomali_enrichment.addMessage(
+                    "ERROR", "An error occurred with the GreyNoise API, please contact GreyNoise for assistance."
+                )
+    except:  # noqa E722
+        anomali_enrichment.addException(
+            "enrichIPSimilar Unknown Error:%sType: %s%sValue:%s"
+            % (os.linesep, sys.exc_info()[0], os.linesep, sys.exc_info()[1])
+        )
+        anomali_enrichment.addMessage(
+            "ERROR", "An Unexpected error occurred, please try again. If the error persists please contact support."
+        )
+    return anomali_enrichment
+
+
+def enrichIPTimeline(anomali_enrichment, search_string):  # noqa: C901
+    try:
+        if api_type.lower() == "community":
+            anomali_enrichment.addMessage("ERROR", "IP Timeline Not Supported with Community API Key")
+        else:
+            # builds response if paid api is being used
+            timeline_response = requests.get(
+                api_base + "v3/noise/ips/" + search_string + "/daily-summary?days=30&limit=50",
+                headers={
+                    "Accept": "application/json",
+                    "key": api_key,
+                    "User-Agent": "greynoise-anomali-enrichment-" + VERSION,
+                },
+            )
+            timeline_response_json = timeline_response.json()
+
+            if (
+                timeline_response.status_code == 200
+                and timeline_response_json
+                and timeline_response_json.get("activity")
+            ):
+                anomali_enrichment.addWidget(
+                    TextWidget(
+                        ItemInWidget(
+                            ItemTypes.String,
+                            "GreyNoise Timeline Details for %s" % search_string,
+                            "GreyNoise Timeline Details for %s" % search_string,
+                            "#A9A9A9",
+                            "#FFFFFF",
+                            "30px",
+                            "bold",
+                        ),
+                        True,
+                    )
+                )
+                anomali_enrichment.addWidget(
+                    TextWidget(
+                        ItemInWidget(
+                            ItemTypes.String,
+                            "Showing Daily Summary of events for last 30 days.  Only days with events will display.",
+                        )
+                    )
+                )
+                anomali_enrichment.addWidget(
+                    TextWidget(
+                        ItemInWidget(
+                            ItemTypes.Link,
+                            "https://viz.greynoise.io/ip/%s?view=timeline" % search_string,
+                            "View IP Timeline Details on GreyNoise Visualizer",
+                        ),
+                        True,
+                    )
+                )
+                # Table Widget #5 Start
+                table_widget_timeline = TableWidget(
+                    "IP Timeline",
+                    [
+                        "Date",
+                        "Classification",
+                        "Tags",
+                        "rDNS",
+                        "Organization",
+                        "ASN",
+                        "Ports",
+                        "Web Paths",
+                        "User Agents",
+                    ],
+                )
+                xrange_data = []
+                asn_dict = {}
+                rdns_dict = {}
+                colors = ["gray", "blue", "purple", "orange", "black"]
+                for activity in reversed(timeline_response_json.get("activity")):
+                    tags = []
+                    for tag in activity["tags"]:
+                        tags.append(tag["name"])
+                    tags_string = ", ".join(tags)
+                    ports = []
+                    for item in activity["protocols"]:
+                        ports.append(str(item["port"]) + "/" + str(item["transport_protocol"]))
+                    ports_string = ", ".join(ports)
+                    paths = ", ".join(activity["http_paths"])
+                    user_agents = "| ".join(activity["http_user_agents"])
+                    date = activity["timestamp"].split("T")[0]
+                    date_object = datetime.datetime.strptime(date, "%Y-%m-%d").date()
+                    date_formatted = date_object.strftime("%-d %b %Y")
+                    table_widget_timeline.addRowOfItems(
+                        [
+                            ItemInWidget(itemValue=date_formatted),
+                            ItemInWidget(itemValue=activity["classification"]),
+                            ItemInWidget(itemValue=tags_string),
+                            ItemInWidget(itemValue=activity["rdns"]),
+                            ItemInWidget(itemValue=activity["organization"]),
+                            ItemInWidget(itemValue=activity["asn"]),
+                            ItemInWidget(itemValue=ports_string),
+                            ItemInWidget(itemValue=paths),
+                            ItemInWidget(itemValue=user_agents),
+                        ]
+                    )
+                    start_date = datetime.datetime.strptime(date, "%Y-%m-%d")
+                    end_date = start_date + datetime.timedelta(days=1)
+                    start_date = int(start_date.timestamp()) * 1000
+                    end_date = int(end_date.timestamp()) * 1000
+                    if activity["classification"]:
+                        if activity["classification"] == "benign":
+                            color = "green"
+                        elif activity["classification"] == "malicious":
+                            color = "red"
+                        else:
+                            color = "gray"
+                        classification_data = {
+                            "x": start_date,
+                            "x2": end_date,
+                            "y": 0,
+                            "color": color,
+                            "value": "Classification: " + activity["classification"],
+                        }
+                        xrange_data.append(classification_data)
+                    if activity["rdns"]:
+                        if activity["rdns"] not in rdns_dict.keys():
+                            rdns_dict[activity["rdns"]] = colors[len(rdns_dict)]
+                        rdns_data = {
+                            "x": start_date,
+                            "x2": end_date,
+                            "y": 1,
+                            "color": rdns_dict[activity["rdns"]],
+                            "value": "rDNS: " + activity["rdns"],
+                        }
+                        xrange_data.append(rdns_data)
+                    if activity["asn"]:
+                        if activity["asn"] not in asn_dict.keys():
+                            asn_dict[activity["asn"]] = colors[len(asn_dict)]
+                        asn_data = {
+                            "x": start_date,
+                            "x2": end_date,
+                            "y": 2,
+                            "color": asn_dict[activity["asn"]],
+                            "value": "ASN: " + activity["asn"],
+                        }
+                        xrange_data.append(asn_data)
+                anomali_enrichment.addWidget(table_widget_timeline)
+
+                # Table Widget #5 End
+
+                # Create XRange ChartWidget
+                def create_chart_widget(template_dict, data):
+                    graph_dict = copy.deepcopy(template_dict)
+                    # fill in the data
+                    graph_dict["series"][0]["data"] = data
+                    graph_json = json.dumps(graph_dict)
+                    xrange_chart_widget = ChartWidget("IP Timeline", graph_json)
+                    return xrange_chart_widget
+
+                chart_widget = create_chart_widget(xrange_template_dict, xrange_data)
+                anomali_enrichment.addWidget(chart_widget)
+            elif timeline_response.status_code == 403:
+                horizontal_rule_widget = HorizontalRuleWidget()
+                anomali_enrichment.addWidget(horizontal_rule_widget)
+                anomali_enrichment.addWidget(
+                    TextWidget(
+                        ItemInWidget(
+                            ItemTypes.String,
+                            "GreyNoise Timeline Details for %s are not available with your Subscription"
+                            % search_string,
+                            "GreyNoise Timeline Details for %s are not available with your Subscription"
+                            % search_string,
+                            "#A9A9A9",
+                            "#FFFFFF",
+                            "30px",
+                            "bold",
+                        ),
+                        True,
+                    )
+                )
+            elif timeline_response.status_code == 401:
+                anomali_enrichment.addMessage("ERROR", "API Key is Missing, Expired or Incorrect, please verify")
+            elif timeline_response.status_code == 429:
+                anomali_enrichment.addMessage("ERROR", "API Rate-Limit Reached, Please try again tomorrow.")
+            elif timeline_response.status_code == 500:
+                anomali_enrichment.addMessage(
+                    "ERROR", "An error occurred with the GreyNoise API, please contact GreyNoise for assistance."
+                )
+    except:  # noqa E722
+        anomali_enrichment.addException(
+            "enrichIPTimeline Unknown Error:%sType: %s%sValue:%s"
+            % (os.linesep, sys.exc_info()[0], os.linesep, sys.exc_info()[1])
+        )
+        anomali_enrichment.addMessage(
+            "ERROR", "An Unexpected error occurred, please try again. If the error persists please contact support."
+        )
+
+    return anomali_enrichment
+
+
+functions = {
+    "enrichIP": enrichIP,
+    "enrichIPSim": enrichIPSim,
+    "enrichIPTimeline": enrichIPTimeline,
+}
 
 if __name__ == "__main__":
     anomali_enrichment = AnomaliEnrichment()